fix: file tree 수정
