+++ app/src/main/java/com/example/atchi_aos/design_system/atom/Card.kt
@@ -1,0 +1,55 @@
+package com.example.atchi_aos.design_system.atom
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun RoundedCard(
+    color: Color,
+    contents: @Composable () -> Unit
+) {
+    Card(
+        modifier = Modifier.fillMaxSize(),
+        shape = RoundedCornerShape(25.dp),
+        colors = CardDefaults.cardColors(color)
+    ) {
+        contents()
+    }
+}
+
+@Preview
+@Composable
+fun RoundedCardPreview() {
+    Column(modifier = Modifier
+        .height(150.dp)
+        .width(150.dp)) {
+        RoundedCard(color = Color(0xFF7544C6)) {
+            Text(
+                text = "HI",
+                fontSize = 18.sp,
+                fontWeight = FontWeight.Bold,
+                color = Color.Black,
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(16.dp),
+                textAlign = TextAlign.Center
+            )
+        }
+    }
+}
+
