--- app/src/main/java/com/example/atchi_aos/design_system/atom/Sample.java
@@ -1,4 +1,0 @@
-package com.example.atchi_aos.design_system.atom;
-
-public class Sample {
-}
