+++ app/src/main/java/com/example/atchi_aos/design_system/atom/surface.kt
@@ -1,4 +1,4 @@
-package com.example.atchi_aos.ui.components
+package com.example.atchi_aos.design_system.atom
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
