+++ app/src/main/java/com/example/atchi_aos/presentation/home/component/ImageCard.kt
@@ -1,4 +1,4 @@
-package com.example.atchi_aos.ui.components
+package com.example.atchi_aos.presentation.home.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -9,40 +9,22 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.atchi_aos.R
-
-@Composable
-fun RoundedCard(
-    color: Color,
-    contents: @Composable () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxSize(),
-        shape = RoundedCornerShape(25.dp),
-        colors = CardDefaults.cardColors(color)
-    ) {
-        contents()
-    }
-}
 
 @Composable
 fun ImageCard(
@@ -89,27 +71,6 @@ fun ImageCard(
                     )
                 )
             }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun RoundedCardPreview() {
-    Column(modifier = Modifier
-        .height(150.dp)
-        .width(150.dp)) {
-        RoundedCard(color = Color(0xFF7544C6)) {
-            Text(
-                text = "HI",
-                fontSize = 18.sp,
-                fontWeight = FontWeight.Bold,
-                color = Color.Black,
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(16.dp),
-                textAlign = TextAlign.Center
-            )
         }
     }
 }
