--- app/src/main/java/com/example/atchi_aos/presentation/home/view/page/Sample.kt
@@ -1,4 +1,0 @@
-package com.example.atchi_aos.presentation.home.view.page
-
-class Sample {
-}
