+++ app/src/main/java/com/example/atchi_aos/presentation/sign/page/signin.kt
@@ -1,4 +1,4 @@
-package com.example.atchi_aos.ui.sign
+package com.example.atchi_aos.presentation.sign.page
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -20,9 +20,9 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.atchi_aos.ui.components.PurpleButton
-import com.example.atchi_aos.ui.components.PurpleField
-import com.example.atchi_aos.ui.components.PurpleLightField
+import com.example.atchi_aos.design_system.atom.PurpleButton
+import com.example.atchi_aos.design_system.atom.PurpleField
+import com.example.atchi_aos.design_system.atom.PurpleLightField
 
 private val headerHeight = 130.dp
 private val bodyHeight = 340.dp
